+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## _Unreleased_
+
+## v0.1.1 - 2017-02-16
 ### Added
 - Pagination for payments list pages [#15](https://github.com/0rora/0rora/issues/15)
 
@@ -10,7 +12,7 @@ All notable changes to this project will be documented in this file.
 - DB type of payments.units is BIGINT, not NUMERIC. Decimal places are not used.
 
 
-## [0.1.0] - 2017-02-10
+## 0.1.0 - 2017-02-10
 
 ### Added
 - List of historical payments under `Payments > History`. (Unpaginated)